fix: `typing.Any` and `object` is not interchangeable in typing context. Refer to PEP 483 for more information. https://www.python.org/dev/peps/pep-0483/#summary-of-gradual-typing
Variable annotated as of type `Any` has all methods while variable annotated as of type `object` has no method. Variable annotated as of type `Any` is consistent with all types and can be assigned to everywhere, while variable annotated as of type `object` is consistent with no types and can be assigned to nowhere except variable of type `object`.